Issue with Missing windows
Hey Varun,
Thanks for updating the repository so we were able to reproduce the issue. We can see that you were only saving the "windows" part of the snapshot, there's also "monitor" data which was not being saved. Since we didn't have the monitor information, the layout engine was unable to determine where to display the windows. We have changed the code to save the entire snapshot and this has fixed the issue.

Please let us know how you get on and thanks for your patience while we got to the bottom of the issue.
Phil.